pyproject.toml
[tool.poetry]
name = "oda-wd-client"
version = "0.0.19a0"
description = "A library for interacting with Workday from Python"
authors = [
    "Karl Fredrik Haugland <karlfredrik.haugland@oda.com>",
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pydantic = ">=1.10.0,<2.0"
suds-py3 = "^1.4.5.0"
pre-commit = "^3.3.3"
yamllint = "^1.32.0"
python-magic = "^0.4.27"

[tool.poetry.group.dev.dependencies]
ipython = "^7.30.1"
isort = "^5.12.0"
mypy = "^1.4.1"
flake8 = "^6.0.0"
black = "^23.7.0"

[tool.black]
target_version = ['py311']
include = '\.pyi?$'
safe = true
exclude = '''
/(
    \.git
  | __pycache__
  | \.venv
)/
'''

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
group_by_package = true
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"

[tool.mypy]
ignore_missing_imports = true
plugins = [
    "pydantic.mypy"
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
